+++ src/gui/globalHotKeys/X11KeyHandler.cpp
@@ -17,14 +17,17 @@
  * Boston, MA 02110-1301, USA.
  */
 
+/*
+ * Inspired by Skycoder42`s QHotKey implementation https://github.com/Skycoder42/QHotkey/blob/master/QHotkey/qhotkey_x11.cpp
+ */
+
 #include "X11KeyHandler.h"
 
-#include <xcb/xcb.h>
-#include <X11/Xlib.h>
+#include "X11ErrorLogger.h"
 
-X11KeyHandler::X11KeyHandler()
+X11KeyHandler::X11KeyHandler() :
+	mFixedModifiers({ 0, Mod2Mask, LockMask, (Mod2Mask | LockMask)})
 {
-	mFixedModifiers = { 0, Mod2Mask, LockMask, (Mod2Mask | LockMask)};
 }
 
 X11KeyHandler::~X11KeyHandler()
@@ -39,9 +42,10 @@ bool X11KeyHandler::registerKey(const QKeySequence &keySequence)
 		return false;
 	}
 
+	X11ErrorLogger x11ErrorLogger;
 	mKeyCodeCombo = mKeyCodeMapper.map(keySequence);
 	for(auto fixedModifier : mFixedModifiers) {
-		XGrabKey(display, mKeyCodeCombo.key, mKeyCodeCombo.modifier | fixedModifier, DefaultRootWindow(display), true, GrabModeAsync, GrabModeAsync);
+		GrabKey(display, fixedModifier);
 	}
 
 	XSync(display, False);
@@ -53,10 +57,18 @@ bool X11KeyHandler::isKeyPressed(void *message)
 	auto genericEvent = static_cast<xcb_generic_event_t *>(message);
 	if (genericEvent->response_type == XCB_KEY_PRESS) {
 		auto keyEvent = static_cast<xcb_key_press_event_t *>(message);
-		return keyEvent->detail == mKeyCodeCombo.key && keyEvent->state == mKeyCodeCombo.modifier;
+
+		for(auto fixedModifier : mFixedModifiers) {
+			if(isMatching(keyEvent, fixedModifier)) {
+				return true;
+			}
+		}
 	}
 	return false;
 }
+
+bool X11KeyHandler::isMatching(const xcb_key_press_event_t *keyEvent, unsigned int fixedModifier) const
+{ return keyEvent->detail == mKeyCodeCombo.key && keyEvent->state == (mKeyCodeCombo.modifier | fixedModifier); }
 
 void X11KeyHandler::unregisterKey() const
 {
@@ -65,9 +77,20 @@ void X11KeyHandler::unregisterKey() const
 		return;
 	}
 
+	X11ErrorLogger x11ErrorLogger;
 	for(auto fixedModifier : mFixedModifiers) {
-		XUngrabKey(display, mKeyCodeCombo.key, mKeyCodeCombo.modifier | fixedModifier, DefaultRootWindow(display));
+		UngrabKey(display, fixedModifier);
 	}
 
 	XSync(display, False);
+}
+
+void X11KeyHandler::GrabKey(Display *display, unsigned int fixedModifier) const
+{
+	XGrabKey(display, mKeyCodeCombo.key, mKeyCodeCombo.modifier | fixedModifier, DefaultRootWindow(display), true, GrabModeAsync, GrabModeAsync);
+}
+
+void X11KeyHandler::UngrabKey(Display *display, unsigned int fixedModifier) const
+{
+	XUngrabKey(display, mKeyCodeCombo.key, mKeyCodeCombo.modifier | fixedModifier, DefaultRootWindow(display));
 }